Adds the page I downloaded from the Wayback Machine
This commit adds the version of hopr.gov.et available on the Wayback
Machine. I haven't tested it to see if it works correctly, what pages
are missing etc. but it's definitely an overkill since it include
multiple entries for the same pages with different JSESSIONID values
(and probably other issues).

It will have to do until the site is live again.